# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: YouTube AI Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /content:
    get:
      operationId: listContent
      tags: [content]
      summary: List generated content
      parameters:
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of content
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Content"
    post:
      operationId: createContent
      tags: [content]
      summary: Generate AI content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentInput"
      responses:
        "201":
          description: Content created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Content"
  /content/{id}:
    get:
      operationId: getContent
      tags: [content]
      summary: Get content by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Content"
        "404":
          description: Not found
    patch:
      operationId: updateContent
      tags: [content]
      summary: Update content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentUpdate"
      responses:
        "200":
          description: Updated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Content"
    delete:
      operationId: deleteContent
      tags: [content]
      summary: Delete content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  # ── VIDEOS ──
...
  /automation/jobs:
    get:
      operationId: listAutomationJobs
      tags: [automation]
      summary: List automation jobs
      parameters:
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Automation jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutomationJob"
    post:
      operationId: createAutomationJob
      tags: [automation]
      summary: Create an automation job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationJobInput"
      responses:
        "201":
          description: Automation job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationJob"
  /automation/jobs/{id}:
    get:
      operationId: getAutomationJob
      tags: [automation]
      summary: Get automation job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Automation job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationJob"
    patch:
      operationId: updateAutomationJob
      tags: [automation]
      summary: Update automation job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationJobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationJob"
    delete:
      operationId: deleteAutomationJob
      tags: [automation]
      summary: Delete automation job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /automation/logs:
    get:
      operationId: getAutomationLogs
      tags: [automation]
      summary: Get automation logs
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutomationLog"

  # ── API SETTINGS ──
...
  /api-settings/{provider}:
    put:
      operationId: upsertApiSetting
      tags: [api-settings]
      summary: Create or update API setting for a provider
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiSettingInput"
      responses:
        "200":
          description: API setting saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSetting"
components:
  schemas:
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    # ── DASHBOARD SCHEMAS ──
...
    Content:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
        topic:
          type: string
        hook:
          type: ["string", "null"]
        outline:
          type: ["string", "null"]
        storyFlow:
          type: ["string", "null"]
        script:
          type: ["string", "null"]
        cta:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        seoTitle:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        hashtags:
          type: ["string", "null"]
        thumbnailPrompt:
          type: ["string", "null"]
        imagePrompt:
          type: ["string", "null"]
        scenePrompt:
          type: ["string", "null"]
        status:
          type: string
        researchId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, type, topic, status, createdAt, updatedAt]
    ContentInput:
      type: object
      properties:
        topic:
          type: string
          minLength: 2
        type:
          type: string
        researchId:
          type: ["integer", "null"]
      required: [topic, type]
    ContentUpdate:
      type: object
      properties:
        title:
          type: string
        hook:
          type: string
        outline:
          type: string
        storyFlow:
          type: string
        script:
          type: string
        cta:
          type: string
        description:
          type: string
        seoTitle:
          type: string
        tags:
          type: string
        hashtags:
          type: string
        thumbnailPrompt:
          type: string
        imagePrompt:
          type: string
        scenePrompt:
          type: string
        status:
          type: string

    # ── VIDEO SCHEMAS ──
...
    AutomationJob:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        status:
          type: string
        frequency:
          type: string
        lastRun:
          type: ["string", "null"]
        nextRun:
          type: ["string", "null"]
        successCount:
          type: integer
        failureCount:
          type: integer
        config:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, type, status, frequency, successCount, failureCount, createdAt]
    AutomationJobInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        type:
          type: string
        frequency:
          type: string
        config:
          type: ["string", "null"]
      required: [name, type, frequency]
    AutomationJobUpdate:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        frequency:
          type: string
        config:
          type: string
    AutomationLog:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        jobName:
          type: string
        status:
          type: string
        message:
          type: string
        createdAt:
          type: string
      required: [id, jobId, jobName, status, message, createdAt]

    # ── API SETTINGS SCHEMAS ──
    ApiSetting:
      type: object
      properties:
        id:
          type: integer
        provider:
          type: string
        label:
          type: string
        isConfigured:
          type: boolean
        description:
          type: ["string", "null"]
        updatedAt:
          type: string
      required: [id, provider, label, isConfigured, updatedAt]
    ApiSettingInput:
      type: object
      properties:
        apiKey:
          type: string
        label:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
      required: [apiKey]