# lib/api-zod/src/generated/types/automationJobInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * YouTube AI Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface AutomationJobInput {
  /** @minLength 2 */
  name: string;
  type: string;
  frequency: string;
  /** @nullable */
  config?: string | null;
}